partial revert of commit SHA-1: 775d010700bc8bb752f8db3753e9e7246093758f :: Inline Debug and Abort
Do NOT inline Debug and Abort functions so we can easily place debugger breakpoints in them (in MSVC2019 debugger); when these functions are inlined, placing breakpoints and *keeping them* is much harder. Besides, the code flow doesn't benefit in run-time performance when inlining these, so there's no loss, as far as I'm concerned.