relax rendering intent check
Closes #1973.
Because there appear to be images with faulty rendering intent in their
ICC profiles in the wild, we relax the check for valid rendering intent
to a warning if we are able to deduce a plausible value for the
rendering for the last byte of the rendering intent of the ICC profiles,
i.e. the byte at offset 67, even if the other bytes (at offsets 64, 65
and 66) are non-zero. In that case we print a warning.